requirements.txt
feedparser==6.0.12
openai==2.24.0
python-dotenv==1.2.1
python_dateutil==2.9.0.post0
Requests==2.32.5
trafilatura==2.0.0
